+++ src/django/api/views/stripe/download_locations_checkout_webhook_view.py
@@ -42,7 +42,7 @@ class DownloadLocationsCheckoutWebhookView(View):
                 payment_id = session["payment_intent"]
                 amount_subtotal = session["amount_subtotal"]
                 amount_total = session["amount_total"]
-                promotion_code = session.get("promotion_code", '')
+                discounts = session["discounts"]
 
                 payment = DownloadLocationPayment(
                     user_id=user_id,
@@ -50,7 +50,7 @@ class DownloadLocationsCheckoutWebhookView(View):
                     payment_id=payment_id,
                     amount_subtotal=amount_subtotal,
                     amount_total=amount_total,
-                    promotion_code=promotion_code,
+                    discounts=discounts,
                 )
                 payment.save()